refactor(test): samle tester i test/unit og test/integration
Flytter Vitest-enhetstester fra src/ til test/unit/ med speilet
mappestruktur, og Playwright-specs fra e2e/ til test/integration/.

Endringer:
- test/unit/**: enhetstester importerer via @src/*-alias
- test/integration/**: Playwright-specs med page objects
- vitest.config.ts: include → test/unit/**/*.test.{ts,tsx}
- playwright.config.ts: testDir → ./test/integration + testMatch
